cleaned dependencies and removed old graphing function
