+++ rating.py
@@ -2,19 +2,12 @@ import pandas as pd
 import numpy as np
 from plotly.subplots import make_subplots
 from scipy import stats, interpolate
-from sklearn.linear_model import LinearRegression
-from scipy.interpolate import interp1d, UnivariateSpline
-from sklearn.preprocessing import StandardScaler
-from sklearn.neural_network import MLPRegressor
+
 import configparser
 import pyodbc
-import configparser
-import pyodbc
-import configparser
 from sqlalchemy import create_engine
 
 import pandas as pd
-import configparser
 
 from datetime import datetime
 from datetime import timedelta
