+++ sql_alchemy.py
@@ -1,29 +1,11 @@
 import pandas as pd
 import numpy as np
 from plotly.subplots import make_subplots
-from scipy import stats, interpolate
-from sklearn.linear_model import LinearRegression
-from scipy.interpolate import interp1d, UnivariateSpline
-from sklearn.preprocessing import StandardScaler
-from sklearn.neural_network import MLPRegressor
-import configparser
-import pyodbc
-import configparser
-import pyodbc
 import configparser
 from sqlalchemy import create_engine
-import psycopg2
-import pandas as pd
-import configparser
-import sys
-import os
-from datetime import datetime
-from datetime import timedelta
 import urllib
-import mysql
 from sqlalchemy import create_engine, Column, Integer, String
 from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import sessionmaker
 
 config = configparser.ConfigParser()
 config.read('gdata_config.ini')
