+++ workup_notes.py
@@ -3,7 +3,6 @@ from datetime import datetime
 from datetime import timedelta
 import urllib
 import configparser
-import time
 import numpy as np
 #import win32com.client as win32
 #import schedule
